% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_particle.R
\name{add_particle}
\alias{add_particle}
\title{Addition of a single sand particle at a random site of the square lattice}
\usage{
add_particle(lattice)
}
\arguments{
\item{lattice}{The configured lattice}
}
\description{
This function randomly selects a site from the square lattice and adds a sand particle to that particular site.
}
\examples{
init(5)
lattice <- config3(lattice)
lattice
lattice <- add_particle(lattice)
lattice

}
