% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config1.R
\name{config1}
\alias{config1}
\title{Configure the initial state of the square lattice}
\usage{
config1(lattice)
}
\arguments{
\item{lattice}{The lattice that has been initially built}
}
\description{
Initializes the lattice, keeping all its elements 0
}
\examples{
init(50)
lattice <- config1(lattice)
lattice

}
