% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config2.R
\name{config2}
\alias{config2}
\title{Configure the initial state of the square lattice}
\usage{
config2(lattice)
}
\arguments{
\item{lattice}{The lattice that has been initially built}
}
\description{
Initializes the lattice, setting all its elements to 1
}
\examples{
init(50)
lattice <- config2(lattice)
lattice

}
