% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config3.R
\name{config3}
\alias{config3}
\title{Configure the initial state of the square lattice}
\usage{
config3(Lattice)
}
\arguments{
\item{Lattice}{The lattice that has been initially built}
}
\description{
Initializes the lattice, setting all its elements to be random integers from 0 to 4
}
\examples{
init(50)
lattice <- config3(lattice)
lattice

}
