% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config4.R
\name{config4}
\alias{config4}
\title{Configure the initial state of the square lattice}
\usage{
config4(Lattice)
}
\arguments{
\item{Lattice}{The lattice that has been initially built}
}
\description{
Initializes the lattice, setting all its elements to be the critical value, 4
}
\examples{
init(50)
lattice <- config4(lattice)
lattice

}
