% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_run.R
\name{event_run}
\alias{event_run}
\title{Main simulation of the Abelian Sandpile (BTW) model}
\usage{
event_run(Lattice, n, max_topplings)
}
\arguments{
\item{Lattice}{A lattice state}

\item{n}{number of sand particles added to the sytem}

\item{max_topplings}{number of maximum topplings allowed in the simulation}
}
\description{
This function runs the main simulation of the BTW model by taking in a particular lattice state, the total number of sand particles to be added in the simulation and the maximum number of topplings allowed. The function then plots:
\enumerate{
\item Average number of sands in the system vs No. of sands added to the system
\item No. of avalanches vs No. of sands added to the system
\item The cumulative probability Prob(s) that an avalanche is of size greater than or equal to s, as a function of s for a lattice of a particular L value
}
}
\examples{
init(50)
lattice <- config3(lattice)
event_run(lattice, 2000, 10000)

}
\references{
\url{http://indico.ictp.it/event/8644/material/3/0.pdf}\cr
\url{https://en.wikipedia.org/wiki/Abelian_sandpile_model}\cr
}
