% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init}
\alias{init}
\title{Initialization}
\usage{
init(L)
}
\arguments{
\item{L}{an integer value, length of the square lattice}
}
\description{
Builds a square lattice, which is a L X L matrix with all elements set to 0
}
\examples{
init(70)

}
